% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VEGAStools.R
\name{annotateVEGAS}
\alias{annotateVEGAS}
\title{Annotate a VEGAS genelist}
\usage{
annotateVEGAS(x, anndb)
}
\arguments{
\item{x}{A gene list of class \code{VEGAS}}

\item{anndb}{An annotation database for an organism of class \code{\link{OrgDb-class}},
available from Bioconductor. The default is \code{\link{org.Hs.eg.db}}.}
}
\value{
If a valid annotation database is specified, an object of class
\code{annVEGAS} which inherits from class \code{VEGAS} and has two extra
columns (\code{Description} and \code{Entrez}). If no valid annotation
database is specified or the default database is not available, the original
\code{VEGAS} object and a warning are returned.
}
\description{
Adds gene name and Entrez identifier for genes in a VEGAS gene list
}
\details{
This only works if an annotation database for the organism under study
is locally available. Also, current matching of gene names does not account
for aliases and does not work terribly well (FIXME).
}

