% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VEGAStools.R
\name{counts}
\alias{counts}
\title{Display expected and observed counts of co-significant genes}
\usage{
counts(x, y, co = 0.05)
}
\arguments{
\item{x,y}{Two gene lists of class \code{VEGAS}}

\item{co}{Cutoff for selecting statistically significant genes in both lists}
}
\value{
A named vector with four components: \code{Obs}, the number of genes
observed to be statistically significant at the chosen cutoff, \code{Exp}, the
number of genes expected to be significant in both lists under the assumption
that the propoerty being statistically significant is independent between the
two lists, and \code{LCL} and \code{UCL}, the respective lower and upper
95\% confidence limits
}
\description{
Given two gene lists of class \code{VEGAS} and a p-value threshold, this
function reports the number of genes below this threshold on both lists,
together with the expected counts under the null hypothesis of no association
between the gene lists, and a 95\% confidence interval
}

