% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VEGAStools.R
\name{exportVEGAS}
\alias{exportVEGAS}
\title{Write an annotated VEGAS genelist to an HTML file}
\usage{
exportVEGAS(x, filename, title, nmax = Inf, co = 1)
}
\arguments{
\item{x}{A gene list of class \code{VEGAS}}

\item{filename}{The name of the output file}

\item{title}{Title of the HTML page}

\item{nmax,co}{Maximum number of genes in output list and p-value cutoff, see
\code{\link{topTable}}}
}
\value{
This function is called for the side effect of generating an HTML
output file
}
\description{
This function combines the functionality of \code{\link{topTable}} and
\code{\link{annotateVEGAS}} to generate annotated lists of top significant
genes with links to the NCBI Entrez database and writes the resulting lists
to a local HTML file.
}
\seealso{
\code{\link[annotate]{htmlpage}}
}

