% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VEGAStools.R
\name{getSharedP}
\alias{getSharedP}
\title{Extract p-values from a list of VEGAS objects}
\usage{
getSharedP(x, co = 1, min_cnt = 2, adjusted = FALSE)
}
\arguments{
\item{x}{A list returned by \code{intersectVEGAS}}

\item{co}{Upper cutoff for p-values to be included}

\item{min_cnt}{Minimum number of VEGAS data sets that have to have
a p-value smaller or equal to \code{co} for a gene to be included}

\item{adjusted}{Logical flag indicating whether to return adjusted
p-values or unadjusted ones (default).}
}
\value{
A matrix of p-values, possibly empty
}
\description{
This function takes the output from \code{} and
returns a matrix of p-values
}
\seealso{
\code{\link{intersectVEGAS}}
}

