% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VEGAStools.R
\name{intersectVEGAS}
\alias{intersectVEGAS}
\title{Intersect VEGAS results}
\usage{
intersectVEGAS(...)
}
\arguments{
\item{...}{A list of VEGAS objects, separated by commas}
}
\value{
A list of VEGAS objects
}
\description{
Takes any number of VEGAS data objects and returns a list with VEGAS data sets
reduced to the shared set of genes
}

