% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VEGAStools.R
\name{summary.VEGAS}
\alias{summary.VEGAS}
\title{Summarize VEGAS data}
\usage{
\method{summary}{VEGAS}(x, ...)
}
\arguments{
\item{x}{A VEGAS object}

\item{...}{Other arguments, currently ignored}
}
\value{
The same object, invisibly
}
\description{
Display a short summary of a VEGAS dataset
}

