% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VEGAStools.R
\name{counts}
\alias{counts}
\alias{plotCounts}
\title{Display expected and observed counts of co-significant genes}
\usage{
counts(x, y, co = 0.05, adjusted = FALSE)

plotCounts(x, y, minP = 1e-06, maxP = 0.1, nP = 100, adjusted = FALSE,
  legend = TRUE, ylim, title, xlab, ylab, ...)
}
\arguments{
\item{x,y}{Two gene lists of class \code{VEGAS}}

\item{co}{Cutoff for selecting statistically significant genes in both lists}

\item{adjusted}{Logical flag indicating whether to use the raw or adjusted p-values (default: \code{FALSE}).}

\item{minP,maxP}{Smallest and largest cutoff value for considering the
overlap in significant genes between lists}

\item{nP}{Number of intermediate points between \code{minP} and \code{maxP}}

\item{legend}{Logical flag indicating whether to add a legend to the plot}

\item{ylim,title,xlab,ylab,...}{Standard graphical prarameters for \code{plot}
                  to override the function defaults}
}
\value{
\code{counts} returns a named vector with five components: \code{Obs},
the number of genes observed to be statistically significant on both lists at the chosen
cutoff, \code{LCL} and \code{UCL}, the corresponding confidence interval
for the number of overlapping genes, \code{Exp}, the number of genes
expected to be significant on both lists under the assumption of no
association between the underlying traits, and \code{p.value}, the
corresponding two-sided p-value.

\code{plotCounts} generates a plot of observed/expected counts as a
function of the specified cutoff range and returns invisibly a data frame
with the cutoffs (column \code{co}) and the corresponding output from
\code{counts}.
}
\description{
Given two gene lists of class \code{VEGAS} and a p-value threshold, this
function reports the number of genes below this threshold on both lists,
with a 95\% confidence interval for the true value, as well as the
the expected number under the null hypothesis of no association
between the gene lists, and the corresponding p-value
}
\details{
Confidence intervals, expected count and p-value are based
on \code{binom.test}. The expected count under the null hypothesis
is conditional on the p-values observed for each trait individually,
i.e. we assume that the probability to be significant on both lists is
the product of the marginal proportions of significant genes on each
list.
\strong{Warning:} inference assumes independence between genes, which is a rather
strong assumption in this setting.
}

