% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VEGAStools.R
\name{readVEGAS}
\alias{readVEGAS}
\title{Read a VEGAS gene list}
\usage{
readVEGAS(fn, path = ".", adjPermP = TRUE)
}
\arguments{
\item{fn}{Name of the file tor read}

\item{path}{Path to file, defaults to \code{.}}

\item{adjPermP}{FLag indicating whether to adjust the permutation p-values
       reported by VEGAS, see below. Defaults to \code{TRUE}.}
}
\value{
A data frame containing the data with extra class attribute \code{VEGAS}
}
\description{
Reads a text file containing output generated by the VEGAS
software
}
\details{
VEGAS returns permutation p-values that are exactly zero, which is
of course impossible. The read-in function adjusts this by including the
observed test statistic in the null-count.
}

